[CI] Fix: version should rely on the branch name.
If we have version or branch name stored into files, some unwanted version/branch
values into these files may appear after a merge, and so CI will deploy
the documentation to the wrong place. To avoid this issue, we should rely
on CI predefined variables indicating the branch it build.